requirements.txt
python>=3.9.7
pytorch>=1.7.1
numpy>=1.21.5
scikit-learn>=1.0.1
scipy>=1.7.3
